CCO.Imp.AG

Documentation

type Ident

Arguments

= String
Type of identifiers.

type Env = [(Ident, Sym)]

An environment maps identifiers to symbol descriptors.

data Sym

A symbol descriptor either describes a variable or a function symbol. For a variable, we store its offset relative to the mark pointer; for a function we store its begin label.

Constructors

V Int

F Int

vars :: [(t, Sym)] -> [(t, Sym)]

Restrict environments to a particular type of descriptor.

funs :: [(t, Sym)] -> [(t, Sym)]

params :: [(t, Sym)] -> [(t, Sym)]

type Syms = [Env]

A symbol table contains descriptors for each variable that is in scope at a certain program point. The table consists of levels of Envs reflecting the nesting of functions: * The head of the list of Envs corresponds to the innermost function scope and so it contains the descriptors of the symbols that are allocated relative to the current mark pointer. * To access the symbols described in the tail of the list, static links are to be followed.

(<<$) :: a -> (a -> b) -> b

Reversed function application operator. Useful in combination with arrow notation: this can be used to feed an argument to multiple functions combined with >>>.

composeList :: [a -> a] -> a -> a

Folds function composition over a list. Useful foir sequencing CodeS.

enterparams :: Env -> CodeS

Produces code for annotating parameters.

enter :: Env -> CodeS

Produces code for entering a block.

exit :: Env -> CodeS

Produces code for exiting a block.

getLink :: Ident -> Syms -> Maybe (CodeS, Int)

If a variable is to be retrieved through a static link, this will produce code retrieving the link address, along with the number of scopes that had to be walked in order to find the SL.

get :: Ident -> Syms -> CodeS

Produces code for retrieving the value of a variable.

withCached :: (Int -> CodeS) -> Ident -> Syms -> CodeS

Generalizes over getCached and setCached. Provide code that either loads or stores a variable at a given offset from the address on the stack.

getCached :: Ident -> Syms -> CodeS

Fetches the value of a global variable through the static-link cache.

setCached :: Ident -> Syms -> CodeS

Sets a global variable value through the static-link cache.

set :: Ident -> Syms -> CodeS

Produces code for assigning a new value to a variable.

call :: Ident -> Syms -> CodeS

Produces code for calling a function.

callGlobal :: Ident -> Syms -> CodeS

Produces code for calling a global function.

return_ :: Syms -> CodeS

Produces code for returning from a function.

wrapSLCacheLoader :: Syms -> CodeS -> CodeS

Preprends a SL cache loader and appends an unloader around some instructions.

loadSLCache :: Syms -> (CodeS, Int)

Provides code for loading the static links that are neccessary for allowing fast access to all relevant variables. Also returns the size of the static link cache.

unloadSLCache :: Int -> CodeS

Given a static link cache size. Pop its contents, which should end one above the current SP. Afterwards, the value the SP pointed to is restored.

cacheVars :: Env -> Syms -> [Ident]

Based on local and global environments, this gives a list of non-local variables that may be used and should therefore be accessible through the static link cache.

data Decl

Constructors

FunDecl Ident [Ident] Stmts

VarDecl Ident

sem_Decl :: Decl -> T_Decl

type T_Decl = [Label] -> Int -> Syms -> (CodeS, Env, [Label], Int)

data Inh_Decl

Constructors

Inh_Decl

Fields

labels_Inh_Decl :: [Label]

offset_Inh_Decl :: Int

syms_Inh_Decl :: Syms

data Syn_Decl

Constructors

Syn_Decl

Fields

codes_Syn_Decl :: CodeS

env_Syn_Decl :: Env

labels_Syn_Decl :: [Label]

offset_Syn_Decl :: Int

wrap_Decl :: T_Decl -> Inh_Decl -> Syn_Decl

sem_Decl_FunDecl :: Ident -> [Ident] -> T_Stmts -> T_Decl

sem_Decl_VarDecl :: Ident -> T_Decl

type Decls = [Decl]

sem_Decls :: Decls -> T_Decls

type T_Decls = [Label] -> Int -> Syms -> (CodeS, Env, [Label], Int)

data Inh_Decls

Constructors

Inh_Decls

Fields

labels_Inh_Decls :: [Label]

offset_Inh_Decls :: Int

syms_Inh_Decls :: Syms

data Syn_Decls

Constructors

Syn_Decls

Fields

codes_Syn_Decls :: CodeS

env_Syn_Decls :: Env

labels_Syn_Decls :: [Label]

offset_Syn_Decls :: Int

wrap_Decls :: T_Decls -> Inh_Decls -> Syn_Decls

sem_Decls_Cons :: T_Decl -> T_Decls -> T_Decls

sem_Decls_Nil :: T_Decls

data Exp

Constructors

Add Exp Exp

Call Ident Exps

Div Exp Exp

Eq Exp Exp

False_

Gt Exp Exp

Int Int

Lt Exp Exp

Mul Exp Exp

Sub Exp Exp

True_

Var Ident

sem_Exp :: Exp -> T_Exp

type T_Exp = [Label] -> Syms -> (String, CodeS, [Label])

data Inh_Exp

Constructors

Inh_Exp

Fields

labels_Inh_Exp :: [Label]

syms_Inh_Exp :: Syms

data Syn_Exp

Constructors

Syn_Exp

Fields

ann_Syn_Exp :: String

codes_Syn_Exp :: CodeS

labels_Syn_Exp :: [Label]

wrap_Exp :: T_Exp -> Inh_Exp -> Syn_Exp

sem_Exp_Add :: T_Exp -> T_Exp -> T_Exp

sem_Exp_Call :: Ident -> T_Exps -> T_Exp

sem_Exp_Div :: T_Exp -> T_Exp -> T_Exp

sem_Exp_Eq :: T_Exp -> T_Exp -> T_Exp

sem_Exp_False_ :: T_Exp

sem_Exp_Gt :: T_Exp -> T_Exp -> T_Exp

sem_Exp_Int :: Int -> T_Exp

sem_Exp_Lt :: T_Exp -> T_Exp -> T_Exp

sem_Exp_Mul :: T_Exp -> T_Exp -> T_Exp

sem_Exp_Sub :: T_Exp -> T_Exp -> T_Exp

sem_Exp_True_ :: T_Exp

sem_Exp_Var :: Ident -> T_Exp

type Exps = [Exp]

sem_Exps :: Exps -> T_Exps

type T_Exps = [Label] -> Syms -> ([String], CodeS, [Label])

data Inh_Exps

Constructors

Inh_Exps

Fields

labels_Inh_Exps :: [Label]

syms_Inh_Exps :: Syms

data Syn_Exps

Constructors

Syn_Exps

Fields

anns_Syn_Exps :: [String]

codes_Syn_Exps :: CodeS

labels_Syn_Exps :: [Label]

wrap_Exps :: T_Exps -> Inh_Exps -> Syn_Exps

sem_Exps_Cons :: T_Exp -> T_Exps -> T_Exps

sem_Exps_Nil :: T_Exps

data Prog

Constructors

TopLevelDecls Decls

sem_Prog :: Prog -> T_Prog

type T_Prog = Code

data Inh_Prog

Constructors

Inh_Prog

data Syn_Prog

Constructors

Syn_Prog

Fields

code_Syn_Prog :: Code

wrap_Prog :: T_Prog -> Inh_Prog -> Syn_Prog

sem_Prog_TopLevelDecls :: T_Decls -> T_Prog

data Stmt

Constructors

Assign Ident Exp

Block Stmts

Call_ Ident Exps

Decl Decl

Empty

If Exp Stmt Stmt

Print Exp

Return Exp

sem_Stmt :: Stmt -> T_Stmt

type T_Stmt = [Label] -> Int -> Syms -> (CodeS, Env, [Label], Int)

data Inh_Stmt

Constructors

Inh_Stmt

Fields

labels_Inh_Stmt :: [Label]

offset_Inh_Stmt :: Int

syms_Inh_Stmt :: Syms

data Syn_Stmt

Constructors

Syn_Stmt

Fields

codes_Syn_Stmt :: CodeS

env_Syn_Stmt :: Env

labels_Syn_Stmt :: [Label]

offset_Syn_Stmt :: Int

wrap_Stmt :: T_Stmt -> Inh_Stmt -> Syn_Stmt

sem_Stmt_Assign :: Ident -> T_Exp -> T_Stmt

sem_Stmt_Block :: T_Stmts -> T_Stmt

sem_Stmt_Call_ :: Ident -> T_Exps -> T_Stmt

sem_Stmt_Decl :: T_Decl -> T_Stmt

sem_Stmt_Empty :: T_Stmt

sem_Stmt_If :: T_Exp -> T_Stmt -> T_Stmt -> T_Stmt

sem_Stmt_Print :: T_Exp -> T_Stmt

sem_Stmt_Return :: T_Exp -> T_Stmt

type Stmts = [Stmt]

sem_Stmts :: Stmts -> T_Stmts

type T_Stmts = [Label] -> Int -> Syms -> (CodeS, Env, [Label], Int)

data Inh_Stmts

Constructors

Inh_Stmts

Fields

labels_Inh_Stmts :: [Label]

offset_Inh_Stmts :: Int

syms_Inh_Stmts :: Syms

data Syn_Stmts

Constructors

Syn_Stmts

Fields

codes_Syn_Stmts :: CodeS

env_Syn_Stmts :: Env

labels_Syn_Stmts :: [Label]

offset_Syn_Stmts :: Int

wrap_Stmts :: T_Stmts -> Inh_Stmts -> Syn_Stmts

sem_Stmts_Cons :: T_Stmt -> T_Stmts -> T_Stmts

sem_Stmts_Nil :: T_Stmts